+++ Gateways/GatewayFactory.php
@@ -14,8 +14,6 @@ final class GatewayFactory
     public static function make(string $gateway, array $config, Options $options): GatewayInterface
     {
         switch ($gateway) {
-            case 'paypay':
-                return new PayPayGateway($config, $options);
             case 'wechat':
                 return new WechatNativeGateway($config, $options);
             case 'alipay':